Linear Parameter models are theoretically complete.
Moved calculating the Moore-Penrose pseudo inverse of
the design matrix into a function call. This can be further
abstracted and optimized. Solving the normal equations is
a function call as well. Any changes would theoretically
only happen at the definition of the design matrix stage,
everything else would be robust.